requirements.txt
asgiref==3.5.2
authy==2.2.6
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-environ==0.9.0
django-phonenumber-field==6.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
phonenumbers==8.12.54
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
twilio==7.13.0
urllib3==1.26.12
